ai-resume-generator/python-backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
python-docx==1.1.2
reportlab==4.2.2
requests==2.32.3
pydantic-settings==2.5.2
